refactor: simulator strategy
