+++ simulator/src/models/config/simulator_factory.py
@@ -1,20 +1,23 @@
 from .sensor_config import SensorConfig
 from ..sensor_type import SensorType
 from ...producers.producer_strategy import ProducerStrategy
-from ...simulators.air_quality_simulator import AirQualitySimulator
-from ...simulators.humidity_simulator import HumiditySimulator
-from ...simulators.parking_simulator import ParkingSimulator
-from ...simulators.recycling_point_simulator import RecyclingPointSimulator
-from ...simulators.simulator import Simulator
-from ...simulators.temperature_simulator import TemperatureSimulator
-from ...simulators.traffic_simulator import TrafficSimulator
-from ...simulators.charging_station_simulator import ChargingStationSimulator
-from ...simulators.precipitation_simulator import PrecipitationSimulator
-from ...simulators.river_level_simulator import RiverLevelSimulator
+from ...simulators.air_quality_simulator_strategy import AirQualitySimulatorStrategy
+from ...simulators.charging_station_simulator_strategy import \
+    ChargingStationSimulatorStrategy
+from ...simulators.humidity_simulator_strategy import HumiditySimulatorStrategy
+from ...simulators.parking_simulator_strategy import ParkingSimulatorStrategy
+from ...simulators.precipitation_simulator_strategy import \
+    PrecipitationSimulatorStrategy
+from ...simulators.recycling_point_simulator_strategy import \
+    RecyclingPointSimulatorStrategy
+from ...simulators.river_level_simulator_strategy import RiverLevelSimulatorStrategy
+from ...simulators.simulator_strategy import SimulatorStrategy
+from ...simulators.temperature_simulator_strategy import TemperatureSimulatorStrategy
+from ...simulators.traffic_simulator_strategy import TrafficSimulatorStrategy
 
 
 def build_simulators(sensors_config: dict[str, any], producer: ProducerStrategy) -> \
-        list[Simulator]:
+        list[SimulatorStrategy]:
     return [
         SimulatorFactory.generate(name, SensorConfig(config), producer) for name, config
         in sensors_config.items()
@@ -24,28 +27,28 @@ def build_simulators(sensors_config: dict[str, any], producer: ProducerStrategy)
 class SimulatorFactory:
     @staticmethod
     def generate(name: str, config: SensorConfig,
-                 producer: ProducerStrategy) -> Simulator:
+                 producer: ProducerStrategy) -> SimulatorStrategy:
         return _get_simulator(name, config, producer)
 
 
 def _get_simulator(name: str, config: SensorConfig,  # noqa: PLR0911
-                   producer: ProducerStrategy) -> Simulator:
+                   producer: ProducerStrategy) -> SimulatorStrategy:
     match config.type:
         case SensorType.AIR_QUALITY:
-            return AirQualitySimulator(name, config, producer)
+            return AirQualitySimulatorStrategy(name, config, producer)
         case SensorType.RECYCLING_POINT:
-            return RecyclingPointSimulator(name, config, producer)
+            return RecyclingPointSimulatorStrategy(name, config, producer)
         case SensorType.PARKING:
-            return ParkingSimulator(name, config, producer)
+            return ParkingSimulatorStrategy(name, config, producer)
         case SensorType.TEMPERATURE:
-            return TemperatureSimulator(name, config, producer)
+            return TemperatureSimulatorStrategy(name, config, producer)
         case SensorType.TRAFFIC:
-            return TrafficSimulator(name, config, producer)
+            return TrafficSimulatorStrategy(name, config, producer)
         case SensorType.HUMIDITY:
-            return HumiditySimulator(name, config, producer)
+            return HumiditySimulatorStrategy(name, config, producer)
         case SensorType.RIVER_LEVEL:
-            return RiverLevelSimulator(name, config, producer)
+            return RiverLevelSimulatorStrategy(name, config, producer)
         case SensorType.PRECIPITATION:
-            return PrecipitationSimulator(name, config, producer)
+            return PrecipitationSimulatorStrategy(name, config, producer)
         case SensorType.CHARGING_STATION:
-            return ChargingStationSimulator(name, config, producer)
+            return ChargingStationSimulatorStrategy(name, config, producer)
