+++ simulator/src/simulators/air_quality_simulator_strategy.py
@@ -3,13 +3,13 @@ from datetime import datetime
 
 from math import pi, sin
 
-from .simulator import Simulator
+from .simulator_strategy import SimulatorStrategy
 from ..models.config.sensor_config import SensorConfig
 from ..models.raw_data.air_quality_raw_data import AirQualityRawData
 from ..producers.producer_strategy import ProducerStrategy
 
 
-class AirQualitySimulator(Simulator):
+class AirQualitySimulatorStrategy(SimulatorStrategy):
 
     def __init__(self, sensor_name: str, config: SensorConfig,
                  producer: ProducerStrategy) -> None:
