+++ simulator/test/simulators/test_charging_station_simulator_strategy.py
@@ -1,24 +1,25 @@
 import unittest
 from datetime import datetime
-from unittest.mock import MagicMock
+from unittest.mock import patch, Mock, MagicMock
 from uuid import UUID
 
 from simulator.src.models.config.sensor_config import SensorConfig
-from simulator.src.models.raw_data.parking_raw_data import ParkingRawData
-from simulator.src.simulators.parking_simulator import ParkingSimulator
+from simulator.src.models.raw_data.charging_station_raw_data import \
+    ChargingStationRawData
+from simulator.src.simulators.charging_station_simulator_strategy import ChargingStationSimulatorStrategy
 
 
-class TestParkingSimulator(unittest.TestCase):
+class TestChargingStationSimulatorStrategy(unittest.TestCase):
     def setUp(self) -> None:
         self.producer = MagicMock()
 
     def test_empty_sensor_name(self) -> None:
         with self.assertRaises(ValueError):
-            ParkingSimulator(
+            ChargingStationSimulatorStrategy(
                 sensor_name='',
                 config=SensorConfig({
                     'uuid': '00000000-0000-0000-0000-000000000000',
-                    'type': 'recycling_point',
+                    'type': 'charging_station',
                     'points_spacing': 'PT1H',
                     'generation_delay': 'PT1H',
                     'latitude': 0,
@@ -27,61 +28,63 @@ class TestParkingSimulator(unittest.TestCase):
                 producer=self.producer,
             )
 
-    @unittest.mock.patch(
-        'random.random',
-        side_effect=[0.1, 0.6, 0.3],
-    )
-    @unittest.mock.patch(
-        'simulator.src.simulators.parking_simulator.ParkingSimulator._generate_next_occupancy_change',
-    )
-    def test_data(self, mock_next_change: MagicMock, _: any) -> None:
-        simulator = ParkingSimulator(
+    @patch('random.random', return_value=0)
+    @patch('random.choices', side_effect=[[11], ['car'], ['car'], ['car']])
+    @patch('random.uniform', side_effect=[3, 50, 2, 80, 1, 20])
+    def test_data(self, _: Mock, __: Mock, ___: Mock) -> None:
+        simulator = ChargingStationSimulatorStrategy(
             sensor_name='test',
             config=SensorConfig({
                 'uuid': '00000000-0000-0000-0000-000000000000',
-                'type': 'recycling_point',
-                'begin_date': datetime(2024, 1, 1),
+                'type': 'charging_station',
+                'limit': 3,
                 'points_spacing': 'PT1H',
                 'generation_delay': 'PT0S',
-                'limit': 3,
+                'begin_date': datetime(2024, 1, 1),
                 'latitude': 0,
                 'longitude': 0,
             }),
             producer=self.producer,
         )
 
-        mock_next_change.side_effect = [
-            datetime(2024, 1, 1, 4),
-            datetime(2024, 1, 1, 5),
-            datetime(2024, 1, 1, 6),
-        ]
-
         stream = [simulator.data() for _ in range(3)]
 
         expected = [
-            ParkingRawData(
-                is_occupied=True,
+            ChargingStationRawData(
+                kwh_supplied=0,
+                remaining_charge_time=0,
                 sensor_uuid=UUID('00000000-0000-0000-0000-000000000000'),
                 sensor_name='test',
                 latitude=0,
                 longitude=0,
                 timestamp=datetime(2024, 1, 1, 0, 0, 0),
+                vehicle_type='bike',
+                battery_level=100,
+                elapsed_time=0,
             ),
-            ParkingRawData(
-                is_occupied=False,
+            ChargingStationRawData(
+                kwh_supplied=0,
+                remaining_charge_time=0,
                 sensor_uuid=UUID('00000000-0000-0000-0000-000000000000'),
                 sensor_name='test',
                 latitude=0,
                 longitude=0,
-                timestamp=datetime(2024, 1, 1, 4, 0, 0),
+                timestamp=datetime(2024, 1, 1, 1, 0, 0),
+                vehicle_type='bike',
+                battery_level=100,
+                elapsed_time=3600,
             ),
-            ParkingRawData(
-                is_occupied=True,
+            ChargingStationRawData(
+                kwh_supplied=0,
+                remaining_charge_time=0,
                 sensor_uuid=UUID('00000000-0000-0000-0000-000000000000'),
                 sensor_name='test',
                 latitude=0,
                 longitude=0,
-                timestamp=datetime(2024, 1, 1, 5, 0, 0),
+                timestamp=datetime(2024, 1, 1, 2, 0, 0),
+                vehicle_type='bike',
+                battery_level=100,
+                elapsed_time=7200,
             ),
         ]
