+++ simulator/test/simulators/test_humidity_simulator_strategy.py
@@ -4,22 +4,21 @@ from unittest.mock import patch, Mock, MagicMock
 from uuid import UUID
 
 from simulator.src.models.config.sensor_config import SensorConfig
-from simulator.src.models.raw_data.recycling_point_raw_data import RecyclingPointRawData
-from simulator.src.simulators.recycling_point_simulator import RecyclingPointSimulator
+from simulator.src.models.raw_data.humidity_raw_data import HumidityRawData
+from simulator.src.simulators.humidity_simulator_strategy import HumiditySimulatorStrategy
 
 
-class TestRecyclingPointSimulator(unittest.TestCase):
-
+class TestHumiditySimulatorStrategy(unittest.TestCase):
     def setUp(self) -> None:
         self.producer = MagicMock()
 
     def test_empty_sensor_name(self) -> None:
         with self.assertRaises(ValueError):
-            RecyclingPointSimulator(
+            HumiditySimulatorStrategy(
                 sensor_name='',
                 config=SensorConfig({
                     'uuid': '00000000-0000-0000-0000-000000000000',
-                    'type': 'recycling_point',
+                    'type': 'humidity',
                     'points_spacing': 'PT1H',
                     'generation_delay': 'PT1H',
                     'latitude': 0,
@@ -29,17 +28,16 @@ class TestRecyclingPointSimulator(unittest.TestCase):
             )
 
     @patch('random.uniform', return_value=0)
-    @patch('builtins.max', return_value=0)
-    def test_data(self, _: Mock, __: Mock) -> None:
-        simulator = RecyclingPointSimulator(
+    def test_data(self, _: Mock) -> None:
+        simulator = HumiditySimulatorStrategy(
             sensor_name='test',
             config=SensorConfig({
                 'uuid': '00000000-0000-0000-0000-000000000000',
-                'type': 'recycling_point',
-                'begin_date': datetime(2024, 1, 1),
+                'type': 'humidity',
+                'limit': 3,
                 'points_spacing': 'PT1H',
                 'generation_delay': 'PT0S',
-                'limit': 3,
+                'begin_date': datetime(2024, 1, 1),
                 'latitude': 0,
                 'longitude': 0,
             }),
@@ -49,24 +47,24 @@ class TestRecyclingPointSimulator(unittest.TestCase):
         stream = [simulator.data() for _ in range(3)]
 
         expected = [
-            RecyclingPointRawData(
-                filling=0,
+            HumidityRawData(
+                value=36.141589368060096,
                 sensor_uuid=UUID('00000000-0000-0000-0000-000000000000'),
                 sensor_name='test',
                 latitude=0,
                 longitude=0,
                 timestamp=datetime(2024, 1, 1, 0, 0, 0),
             ),
-            RecyclingPointRawData(
-                filling=0,
+            HumidityRawData(
+                value=39.24770500506624,
                 sensor_uuid=UUID('00000000-0000-0000-0000-000000000000'),
                 sensor_name='test',
                 latitude=0,
                 longitude=0,
                 timestamp=datetime(2024, 1, 1, 1, 0, 0),
             ),
-            RecyclingPointRawData(
-                filling=0,
+            HumidityRawData(
+                value=42.30067414234525,
                 sensor_uuid=UUID('00000000-0000-0000-0000-000000000000'),
                 sensor_name='test',
                 latitude=0,
