+++ simulator/test/simulators/test_parking_simulator_strategy.py
@@ -1,40 +1,45 @@
 import unittest
 from datetime import datetime
-from unittest.mock import patch, MagicMock
+from unittest.mock import MagicMock
 from uuid import UUID
 
 from simulator.src.models.config.sensor_config import SensorConfig
-from simulator.src.models.raw_data.temperature_raw_data import TemperatureRawData
-from simulator.src.simulators.temperature_simulator import TemperatureSimulator
+from simulator.src.models.raw_data.parking_raw_data import ParkingRawData
+from simulator.src.simulators.parking_simulator_strategy import ParkingSimulatorStrategy
 
 
-class TestTemperatureSimulator(unittest.TestCase):
-
+class TestParkingSimulatorStrategy(unittest.TestCase):
     def setUp(self) -> None:
         self.producer = MagicMock()
 
     def test_empty_sensor_name(self) -> None:
         with self.assertRaises(ValueError):
-            TemperatureSimulator(
+            ParkingSimulatorStrategy(
                 sensor_name='',
                 config=SensorConfig({
                     'uuid': '00000000-0000-0000-0000-000000000000',
-                    'type': 'temperature',
-                    'points_spacing': 'PT1S',
-                    'generation_delay': 'PT1S',
+                    'type': 'recycling_point',
+                    'points_spacing': 'PT1H',
+                    'generation_delay': 'PT1H',
                     'latitude': 0,
                     'longitude': 0,
                 }),
                 producer=self.producer,
             )
 
-    @patch('random.uniform', return_value=0)
-    def test_data(self, _: any) -> None:
-        simulator = TemperatureSimulator(
+    @unittest.mock.patch(
+        'random.random',
+        side_effect=[0.1, 0.6, 0.3],
+    )
+    @unittest.mock.patch(
+        'simulator.src.simulators.parking_simulator_strategy.ParkingSimulatorStrategy._generate_next_occupancy_change',
+    )
+    def test_data(self, mock_next_change: MagicMock, _: any) -> None:
+        simulator = ParkingSimulatorStrategy(
             sensor_name='test',
             config=SensorConfig({
                 'uuid': '00000000-0000-0000-0000-000000000000',
-                'type': 'temperature',
+                'type': 'recycling_point',
                 'begin_date': datetime(2024, 1, 1),
                 'points_spacing': 'PT1H',
                 'generation_delay': 'PT0S',
@@ -45,32 +50,38 @@ class TestTemperatureSimulator(unittest.TestCase):
             producer=self.producer,
         )
 
+        mock_next_change.side_effect = [
+            datetime(2024, 1, 1, 4),
+            datetime(2024, 1, 1, 5),
+            datetime(2024, 1, 1, 6),
+        ]
+
         stream = [simulator.data() for _ in range(3)]
 
         expected = [
-            TemperatureRawData(
-                value=3.153388482065103,
+            ParkingRawData(
+                is_occupied=True,
                 sensor_uuid=UUID('00000000-0000-0000-0000-000000000000'),
                 sensor_name='test',
                 latitude=0,
                 longitude=0,
                 timestamp=datetime(2024, 1, 1, 0, 0, 0),
             ),
-            TemperatureRawData(
-                value=4.395834736867558,
+            ParkingRawData(
+                is_occupied=False,
                 sensor_uuid=UUID('00000000-0000-0000-0000-000000000000'),
                 sensor_name='test',
                 latitude=0,
                 longitude=0,
-                timestamp=datetime(2024, 1, 1, 1, 0, 0),
+                timestamp=datetime(2024, 1, 1, 4, 0, 0),
             ),
-            TemperatureRawData(
-                value=5.617022391779162,
+            ParkingRawData(
+                is_occupied=True,
                 sensor_uuid=UUID('00000000-0000-0000-0000-000000000000'),
                 sensor_name='test',
                 latitude=0,
                 longitude=0,
-                timestamp=datetime(2024, 1, 1, 2, 0, 0),
+                timestamp=datetime(2024, 1, 1, 5, 0, 0),
             ),
         ]
