+++ simulator/test/simulators/test_precipitation_simulator_strategy.py
@@ -4,22 +4,21 @@ from unittest.mock import patch, Mock, MagicMock
 from uuid import UUID
 
 from simulator.src.models.config.sensor_config import SensorConfig
-from simulator.src.models.raw_data.charging_station_raw_data import \
-    ChargingStationRawData
-from simulator.src.simulators.charging_station_simulator import ChargingStationSimulator
+from simulator.src.models.raw_data.precipitation_raw_data import PrecipitationRawData
+from simulator.src.simulators.precipitation_simulator_strategy import PrecipitationSimulatorStrategy
 
 
-class TestChargingStationSimulator(unittest.TestCase):
+class TestPrecipitationSimulatorStrategy(unittest.TestCase):
     def setUp(self) -> None:
         self.producer = MagicMock()
 
     def test_empty_sensor_name(self) -> None:
         with self.assertRaises(ValueError):
-            ChargingStationSimulator(
+            PrecipitationSimulatorStrategy(
                 sensor_name='',
                 config=SensorConfig({
                     'uuid': '00000000-0000-0000-0000-000000000000',
-                    'type': 'charging_station',
+                    'type': 'precipitation',
                     'points_spacing': 'PT1H',
                     'generation_delay': 'PT1H',
                     'latitude': 0,
@@ -28,15 +27,14 @@ class TestChargingStationSimulator(unittest.TestCase):
                 producer=self.producer,
             )
 
-    @patch('random.random', return_value=0)
-    @patch('random.choices', side_effect=[[11], ['car'], ['car'], ['car']])
-    @patch('random.uniform', side_effect=[3, 50, 2, 80, 1, 20])
-    def test_data(self, _: Mock, __: Mock, ___: Mock) -> None:
-        simulator = ChargingStationSimulator(
+    @patch('random.uniform', side_effect=[0.8, 0.9, 0.5])
+    @patch('random.random', side_effect=[0.86, 0.88, 0.9, 0.92, 0.94, 0.96, 0.98, 1.0])
+    def test_data(self, _: Mock, __: Mock) -> None:
+        simulator = PrecipitationSimulatorStrategy(
             sensor_name='test',
             config=SensorConfig({
                 'uuid': '00000000-0000-0000-0000-000000000000',
-                'type': 'charging_station',
+                'type': 'precipitation',
                 'limit': 3,
                 'points_spacing': 'PT1H',
                 'generation_delay': 'PT0S',
@@ -50,41 +48,29 @@ class TestChargingStationSimulator(unittest.TestCase):
         stream = [simulator.data() for _ in range(3)]
 
         expected = [
-            ChargingStationRawData(
-                kwh_supplied=0,
-                remaining_charge_time=0,
+            PrecipitationRawData(
+                value=0,
                 sensor_uuid=UUID('00000000-0000-0000-0000-000000000000'),
                 sensor_name='test',
                 latitude=0,
                 longitude=0,
                 timestamp=datetime(2024, 1, 1, 0, 0, 0),
-                vehicle_type='bike',
-                battery_level=100,
-                elapsed_time=0,
             ),
-            ChargingStationRawData(
-                kwh_supplied=0,
-                remaining_charge_time=0,
+            PrecipitationRawData(
+                value=1.755517366541557,
                 sensor_uuid=UUID('00000000-0000-0000-0000-000000000000'),
                 sensor_name='test',
                 latitude=0,
                 longitude=0,
                 timestamp=datetime(2024, 1, 1, 1, 0, 0),
-                vehicle_type='bike',
-                battery_level=100,
-                elapsed_time=3600,
             ),
-            ChargingStationRawData(
-                kwh_supplied=0,
-                remaining_charge_time=0,
+            PrecipitationRawData(
+                value=0.9752874258564206,
                 sensor_uuid=UUID('00000000-0000-0000-0000-000000000000'),
                 sensor_name='test',
                 latitude=0,
                 longitude=0,
                 timestamp=datetime(2024, 1, 1, 2, 0, 0),
-                vehicle_type='bike',
-                battery_level=100,
-                elapsed_time=7200,
             ),
         ]
