+++ simulator/test/simulators/test_recycling_point_simulator_strategy.py
@@ -4,21 +4,22 @@ from unittest.mock import patch, Mock, MagicMock
 from uuid import UUID
 
 from simulator.src.models.config.sensor_config import SensorConfig
-from simulator.src.models.raw_data.traffic_raw_data import TrafficRawData
-from simulator.src.simulators.traffic_simulator import TrafficSimulator
+from simulator.src.models.raw_data.recycling_point_raw_data import RecyclingPointRawData
+from simulator.src.simulators.recycling_point_simulator_strategy import RecyclingPointSimulatorStrategy
 
 
-class TestTrafficSimulator(unittest.TestCase):
+class TestRecyclingPointSimulatorStrategy(unittest.TestCase):
+
     def setUp(self) -> None:
         self.producer = MagicMock()
 
     def test_empty_sensor_name(self) -> None:
         with self.assertRaises(ValueError):
-            TrafficSimulator(
+            RecyclingPointSimulatorStrategy(
                 sensor_name='',
                 config=SensorConfig({
                     'uuid': '00000000-0000-0000-0000-000000000000',
-                    'type': 'traffic',
+                    'type': 'recycling_point',
                     'points_spacing': 'PT1H',
                     'generation_delay': 'PT1H',
                     'latitude': 0,
@@ -28,16 +29,17 @@ class TestTrafficSimulator(unittest.TestCase):
             )
 
     @patch('random.uniform', return_value=0)
-    def test_data(self, _: Mock) -> None:
-        simulator = TrafficSimulator(
+    @patch('builtins.max', return_value=0)
+    def test_data(self, _: Mock, __: Mock) -> None:
+        simulator = RecyclingPointSimulatorStrategy(
             sensor_name='test',
             config=SensorConfig({
                 'uuid': '00000000-0000-0000-0000-000000000000',
-                'type': 'traffic',
-                'limit': 3,
+                'type': 'recycling_point',
+                'begin_date': datetime(2024, 1, 1),
                 'points_spacing': 'PT1H',
                 'generation_delay': 'PT0S',
-                'begin_date': datetime(2024, 1, 1),
+                'limit': 3,
                 'latitude': 0,
                 'longitude': 0,
             }),
@@ -47,27 +49,24 @@ class TestTrafficSimulator(unittest.TestCase):
         stream = [simulator.data() for _ in range(3)]
 
         expected = [
-            TrafficRawData(
-                avg_speed=32.470887891211184,
-                vehicles=39,
+            RecyclingPointRawData(
+                filling=0,
                 sensor_uuid=UUID('00000000-0000-0000-0000-000000000000'),
                 sensor_name='test',
                 latitude=0,
                 longitude=0,
                 timestamp=datetime(2024, 1, 1, 0, 0, 0),
             ),
-            TrafficRawData(
-                avg_speed=34.12195312644378,
-                vehicles=39,
+            RecyclingPointRawData(
+                filling=0,
                 sensor_uuid=UUID('00000000-0000-0000-0000-000000000000'),
                 sensor_name='test',
                 latitude=0,
                 longitude=0,
                 timestamp=datetime(2024, 1, 1, 1, 0, 0),
             ),
-            TrafficRawData(
-                avg_speed=34.08242621271829,
-                vehicles=37,
+            RecyclingPointRawData(
+                filling=0,
                 sensor_uuid=UUID('00000000-0000-0000-0000-000000000000'),
                 sensor_name='test',
                 latitude=0,
