+++ simulator/test/simulators/test_river_level_simulator_strategy.py
@@ -4,21 +4,21 @@ from unittest.mock import patch, Mock, MagicMock
 from uuid import UUID
 
 from simulator.src.models.config.sensor_config import SensorConfig
-from simulator.src.models.raw_data.precipitation_raw_data import PrecipitationRawData
-from simulator.src.simulators.precipitation_simulator import PrecipitationSimulator
+from simulator.src.models.raw_data.river_level_raw_data import RiverLevelRawData
+from simulator.src.simulators.river_level_simulator_strategy import RiverLevelSimulatorStrategy
 
 
-class TestPrecipitationSimulator(unittest.TestCase):
+class TestRiverLevelSimulatorStrategy(unittest.TestCase):
     def setUp(self) -> None:
         self.producer = MagicMock()
 
     def test_empty_sensor_name(self) -> None:
         with self.assertRaises(ValueError):
-            PrecipitationSimulator(
+            RiverLevelSimulatorStrategy(
                 sensor_name='',
                 config=SensorConfig({
                     'uuid': '00000000-0000-0000-0000-000000000000',
-                    'type': 'precipitation',
+                    'type': 'river_level',
                     'points_spacing': 'PT1H',
                     'generation_delay': 'PT1H',
                     'latitude': 0,
@@ -27,14 +27,13 @@ class TestPrecipitationSimulator(unittest.TestCase):
                 producer=self.producer,
             )
 
-    @patch('random.uniform', side_effect=[0.8, 0.9, 0.5])
-    @patch('random.random', side_effect=[0.86, 0.88, 0.9, 0.92, 0.94, 0.96, 0.98, 1.0])
-    def test_data(self, _: Mock, __: Mock) -> None:
-        simulator = PrecipitationSimulator(
+    @patch('random.gauss', side_effect=[0.8, 0.9, 0.5])
+    def test_data(self, _: Mock) -> None:
+        simulator = RiverLevelSimulatorStrategy(
             sensor_name='test',
             config=SensorConfig({
                 'uuid': '00000000-0000-0000-0000-000000000000',
-                'type': 'precipitation',
+                'type': 'river_level',
                 'limit': 3,
                 'points_spacing': 'PT1H',
                 'generation_delay': 'PT0S',
@@ -48,24 +47,24 @@ class TestPrecipitationSimulator(unittest.TestCase):
         stream = [simulator.data() for _ in range(3)]
 
         expected = [
-            PrecipitationRawData(
-                value=0,
+            RiverLevelRawData(
+                value=1608.2567562007653,
                 sensor_uuid=UUID('00000000-0000-0000-0000-000000000000'),
                 sensor_name='test',
                 latitude=0,
                 longitude=0,
                 timestamp=datetime(2024, 1, 1, 0, 0, 0),
             ),
-            PrecipitationRawData(
-                value=1.755517366541557,
+            RiverLevelRawData(
+                value=1809.2888507258608,
                 sensor_uuid=UUID('00000000-0000-0000-0000-000000000000'),
                 sensor_name='test',
                 latitude=0,
                 longitude=0,
                 timestamp=datetime(2024, 1, 1, 1, 0, 0),
             ),
-            PrecipitationRawData(
-                value=0.9752874258564206,
+            RiverLevelRawData(
+                value=1005.1604726254782,
                 sensor_uuid=UUID('00000000-0000-0000-0000-000000000000'),
                 sensor_name='test',
                 latitude=0,
