+++ simulator/test/simulators/test_temperature_simulator_strategy.py
@@ -1,43 +1,44 @@
 import unittest
 from datetime import datetime
-from unittest.mock import patch, Mock, MagicMock
+from unittest.mock import patch, MagicMock
 from uuid import UUID
 
 from simulator.src.models.config.sensor_config import SensorConfig
-from simulator.src.models.raw_data.river_level_raw_data import RiverLevelRawData
-from simulator.src.simulators.river_level_simulator import RiverLevelSimulator
+from simulator.src.models.raw_data.temperature_raw_data import TemperatureRawData
+from simulator.src.simulators.temperature_simulator_strategy import TemperatureSimulatorStrategy
 
 
-class TestRiverLevelSimulator(unittest.TestCase):
+class TestTemperatureSimulatorStrategy(unittest.TestCase):
+
     def setUp(self) -> None:
         self.producer = MagicMock()
 
     def test_empty_sensor_name(self) -> None:
         with self.assertRaises(ValueError):
-            RiverLevelSimulator(
+            TemperatureSimulatorStrategy(
                 sensor_name='',
                 config=SensorConfig({
                     'uuid': '00000000-0000-0000-0000-000000000000',
-                    'type': 'river_level',
-                    'points_spacing': 'PT1H',
-                    'generation_delay': 'PT1H',
+                    'type': 'temperature',
+                    'points_spacing': 'PT1S',
+                    'generation_delay': 'PT1S',
                     'latitude': 0,
                     'longitude': 0,
                 }),
                 producer=self.producer,
             )
 
-    @patch('random.gauss', side_effect=[0.8, 0.9, 0.5])
-    def test_data(self, _: Mock) -> None:
-        simulator = RiverLevelSimulator(
+    @patch('random.uniform', return_value=0)
+    def test_data(self, _: any) -> None:
+        simulator = TemperatureSimulatorStrategy(
             sensor_name='test',
             config=SensorConfig({
                 'uuid': '00000000-0000-0000-0000-000000000000',
-                'type': 'river_level',
-                'limit': 3,
+                'type': 'temperature',
+                'begin_date': datetime(2024, 1, 1),
                 'points_spacing': 'PT1H',
                 'generation_delay': 'PT0S',
-                'begin_date': datetime(2024, 1, 1),
+                'limit': 3,
                 'latitude': 0,
                 'longitude': 0,
             }),
@@ -47,24 +48,24 @@ class TestRiverLevelSimulator(unittest.TestCase):
         stream = [simulator.data() for _ in range(3)]
 
         expected = [
-            RiverLevelRawData(
-                value=1608.2567562007653,
+            TemperatureRawData(
+                value=3.153388482065103,
                 sensor_uuid=UUID('00000000-0000-0000-0000-000000000000'),
                 sensor_name='test',
                 latitude=0,
                 longitude=0,
                 timestamp=datetime(2024, 1, 1, 0, 0, 0),
             ),
-            RiverLevelRawData(
-                value=1809.2888507258608,
+            TemperatureRawData(
+                value=4.395834736867558,
                 sensor_uuid=UUID('00000000-0000-0000-0000-000000000000'),
                 sensor_name='test',
                 latitude=0,
                 longitude=0,
                 timestamp=datetime(2024, 1, 1, 1, 0, 0),
             ),
-            RiverLevelRawData(
-                value=1005.1604726254782,
+            TemperatureRawData(
+                value=5.617022391779162,
                 sensor_uuid=UUID('00000000-0000-0000-0000-000000000000'),
                 sensor_name='test',
                 latitude=0,
