+++ simulator/test/simulators/test_traffic_simulator_strategy.py
@@ -4,21 +4,21 @@ from unittest.mock import patch, Mock, MagicMock
 from uuid import UUID
 
 from simulator.src.models.config.sensor_config import SensorConfig
-from simulator.src.models.raw_data.humidity_raw_data import HumidityRawData
-from simulator.src.simulators.humidity_simulator import HumiditySimulator
+from simulator.src.models.raw_data.traffic_raw_data import TrafficRawData
+from simulator.src.simulators.traffic_simulator_strategy import TrafficSimulatorStrategy
 
 
-class TestHumiditySimulator(unittest.TestCase):
+class TestTrafficSimulatorStrategy(unittest.TestCase):
     def setUp(self) -> None:
         self.producer = MagicMock()
 
     def test_empty_sensor_name(self) -> None:
         with self.assertRaises(ValueError):
-            HumiditySimulator(
+            TrafficSimulatorStrategy(
                 sensor_name='',
                 config=SensorConfig({
                     'uuid': '00000000-0000-0000-0000-000000000000',
-                    'type': 'humidity',
+                    'type': 'traffic',
                     'points_spacing': 'PT1H',
                     'generation_delay': 'PT1H',
                     'latitude': 0,
@@ -29,11 +29,11 @@ class TestHumiditySimulator(unittest.TestCase):
 
     @patch('random.uniform', return_value=0)
     def test_data(self, _: Mock) -> None:
-        simulator = HumiditySimulator(
+        simulator = TrafficSimulatorStrategy(
             sensor_name='test',
             config=SensorConfig({
                 'uuid': '00000000-0000-0000-0000-000000000000',
-                'type': 'humidity',
+                'type': 'traffic',
                 'limit': 3,
                 'points_spacing': 'PT1H',
                 'generation_delay': 'PT0S',
@@ -47,24 +47,27 @@ class TestHumiditySimulator(unittest.TestCase):
         stream = [simulator.data() for _ in range(3)]
 
         expected = [
-            HumidityRawData(
-                value=36.141589368060096,
+            TrafficRawData(
+                avg_speed=32.470887891211184,
+                vehicles=39,
                 sensor_uuid=UUID('00000000-0000-0000-0000-000000000000'),
                 sensor_name='test',
                 latitude=0,
                 longitude=0,
                 timestamp=datetime(2024, 1, 1, 0, 0, 0),
             ),
-            HumidityRawData(
-                value=39.24770500506624,
+            TrafficRawData(
+                avg_speed=34.12195312644378,
+                vehicles=39,
                 sensor_uuid=UUID('00000000-0000-0000-0000-000000000000'),
                 sensor_name='test',
                 latitude=0,
                 longitude=0,
                 timestamp=datetime(2024, 1, 1, 1, 0, 0),
             ),
-            HumidityRawData(
-                value=42.30067414234525,
+            TrafficRawData(
+                avg_speed=34.08242621271829,
+                vehicles=37,
                 sensor_uuid=UUID('00000000-0000-0000-0000-000000000000'),
                 sensor_name='test',
                 latitude=0,
